AI-Stock-Chatbot/user/backened/pythonBackened-AI-Stocks/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
python-dotenv
passlib[bcrypt]==1.7.4
bcrypt==3.2.2